refactor: service, constants, ...
