+++ src/constants.ts
@@ -50,10 +50,18 @@ export const nameMap = Object.fromEntries(
     Object.entries(SpeakerIdMap).map(([k, v]) => [v, k])
 )
 
-const baseSpeakId = 114514
+const baseSpeakId = 114513
 
-export const SpeakerKeyIdMap = Object.entries(SpeakerIdMap).map(
-    ([k, v], index) => [k, baseSpeakId + index + 1]
-)
+// as {key:value}
+export const SpeakerKeyIdMap = Object.entries(SpeakerIdMap)
+    .sort((a, b) => (a[0] < b[0] ? 1 : -1))
+    .map(([k, v], index) => [k, baseSpeakId + index])
+    .reduce(
+        (acc, [k, v]) => {
+            acc[v as number] = k as string
+            return acc
+        },
+        {} as Record<number, string>
+    )
 
 export const Language = ['ZH', 'JP', 'EN', 'AUTO', 'MIX']
