+++ src/types.ts
@@ -1,0 +1,11 @@
+export interface SpeakConfig {
+    speaker: string
+    sdp_ratio: number
+    noise: number
+    noisew: number
+    length: number
+    prompt: string
+    weight: number
+    language: string
+    input: string
+}
